working worm

+++ Assets/Player/DefaultBullet.cs
@@ -1,12 +1,13 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+
 class DefaultBullet : MonoBehaviour
 {
     public float speed = 100f;
-    public float damage = 1f;
+    public int damage = 1;
     private Rigidbody _rb;
-    
+
     void Awake()
     {
         _rb = GetComponent<Rigidbody>();
@@ -19,9 +20,9 @@ class DefaultBullet : MonoBehaviour
             Destroy(this.gameObject);
         }
     }
+
     public void SetDirection(Vector3 direction)
     {
         _rb.velocity = direction * speed;
     }
 }
-
